fix(lint): 修复 golangci-lint v2 报告的 7 个代码问题
- G115: GORM Model ID 类型 uint64 → int64 (与领域层一致)
  消除 int64↔uint64 转换溢出告警（5处）
- goconst: health.go 中 "status" 抽取为常量 statusKey (1处)
- goconst: entity_test.go 测试数据加 nolint 注释 (1处)

+++ internal/infrastructure/persistence/gorm/article_model.go
@@ -6,10 +6,10 @@ import "time"
 // It is intentionally separate from domain/article.Article to avoid leaking
 // ORM concerns into the domain layer.
 type ArticleModel struct {
-	ID         uint64    `gorm:"primaryKey;column:id"`
+	ID         int64    `gorm:"primaryKey;column:id"`
 	Title      string    `gorm:"size:200;not null;column:title"`
 	Content    string    `gorm:"type:text;not null;column:content"`
-	AuthorID   uint64    `gorm:"not null;index;column:author_id"`
+	AuthorID   int64    `gorm:"not null;index;column:author_id"`
 	AuthorName string    `gorm:"size:100;not null;default:'';column:author_name"`
 	CreatedAt  time.Time `gorm:"column:created_at"`
 	UpdatedAt  time.Time `gorm:"column:updated_at"`
@@ -21,8 +21,8 @@ func (ArticleModel) TableName() string { return "articles" }
 // toEntity converts GORM model → domain entity.
 func (m *ArticleModel) toEntity() *articleEntity {
 	return &articleEntity{
-		ID: int64(m.ID), Title: m.Title, Content: m.Content,
-		AuthorID: int64(m.AuthorID), AuthorName: m.AuthorName,
+		ID: m.ID, Title: m.Title, Content: m.Content,
+		AuthorID: m.AuthorID, AuthorName: m.AuthorName,
 		CreatedAt: m.CreatedAt, UpdatedAt: m.UpdatedAt,
 	}
 }
